+++ MakeYouPro.Bourse.Rates.Dal/Models/USDDto.cs
@@ -15,13 +15,13 @@ namespace MakeYouPro.Bourse.Rates.Dal.Models
         [Required]
         public int Id { get; set; }
 
-        public double USDRUB { get; set; }
-        public double USDEUR { get; set; }
-        public double USDJPY { get; set; }
-        public double USDCNY { get; set; }
-        public double USDRSD { get; set; }
-        public double USDBGN { get; set; }
-        public double USDARS { get; set; }
+        public decimal USDRUB { get; set; }
+        public decimal USDEUR { get; set; }
+        public decimal USDJPY { get; set; }
+        public decimal USDCNY { get; set; }
+        public decimal USDRSD { get; set; }
+        public decimal USDBGN { get; set; }
+        public decimal USDARS { get; set; }
 
 
         [Required]
